+++ client/src/App.tsx
@@ -2,12 +2,14 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
+import { CustomCursor } from './components/ui/CustomCursor'
 
 function App() {
   const [count, setCount] = useState(0)
 
   return (
     <>
+      <CustomCursor />
       <div>
         <a href="https://vite.dev" target="_blank">
           <img src={viteLogo} className="logo" alt="Vite logo" />